Reorganize heatmap files
